+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-07-05 11:24:57
-// Commit: test(ui): implement widget integration tests for ui
+// Updated at: 2026-07-05 21:53:18
+// Commit: feat(ui): add bi-directional event bus handler for ui
 
 class uiComponentState {
-  final String commitRef = '2595';
-  final String lastMessage = 'test(ui): implement widget integration tests for ui';
-  final DateTime timestamp = DateTime.parse('2026-07-05T11:24:57+05:00');
+  final String commitRef = '2617';
+  final String lastMessage = 'feat(ui): add bi-directional event bus handler for ui';
+  final DateTime timestamp = DateTime.parse('2026-07-05T21:53:18+05:00');
 }